@@ -11,15 +11,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
+    "@angular/animations": "^7.2.15",
+    "@angular/cdk": "^8.0.1",
     "@angular/common": "~7.2.0",
     "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
+    "@angular/core": "^8.0.2",
     "@angular/forms": "~7.2.0",
+    "@angular/material": "^8.0.1",
     "@angular/platform-browser": "~7.2.0",
     "@angular/platform-browser-dynamic": "~7.2.0",
     "@angular/router": "~7.2.0",
+    "angular": "^1.7.8",
+    "chart.js": "^2.8.0",
     "core-js": "^2.5.4",
+    "ng2-charts": "^2.3.0",
     "ngx-logger": "^3.3.11",
     "ngx-socket-io": "^2.1.1",
     "p5": "^0.7.3",
